Formatting and style

+++ app/src/main/java/com/example/redditwalls/repositories/RWRepository.kt
@@ -21,6 +21,6 @@ class RWRepository @Inject constructor() {
 
     suspend fun searchSubs(query: String) = RWApi.searchSubs(query)
 
-    suspend fun getPostInfo(postLink: String, imageSize: Int) =
+    suspend fun getPostInfo(postLink: String, imageSize: Double) =
         RWApi.getPostInfo(postLink, imageSize)
 }
